Fix local parameter check
Local parameters can't have initial assignments nor assignment rules.  Checking to see if they do actually introduces a subtle bug:  if the model has a global parameter (or species, etc.) with the same ID as a local parameter that *did* have an assignment, this check would find that and mistakenly attribute it to the local parameter.