Fix for #96: consolidate warnings
This is not a general fix for consolidating warnings, but consolidates two sets of warnings, and fixes a third.

* The SBML warnings are now consolidated by error number.
* The ModelChange 'cannot validate XPath' messages are now simply "Model change XPaths are currently not validated."  This is because the code does not even attempt to validate ModelChange XPaths.
* The DataGenerator code now checks to make sure that there aren't *structural* ModelChanges in the referenced model instead of checking to see if there are *any* ModelChanges.  ModelChanges that change attributes or are of type ComputeChange will not invalidate any other XPath in the SED-ML, unlike the AddXML, ReplaceXML, and RemoveXML changes.  Fortunately, the latter are rare to nonexistent.

...except that I now realize as I type this that a ChangeAttribute could change the ID of an element, making an XPath that relied on that attribute invalid.  But honestly, if you do that, and biosimulators_utils tells you that you have an invalid XPath, I think you deserve it.  I can't see it actually happening in the wild.

However, I do see the danger, and perhaps the goal eventually was to *apply* the ModelChanges before validating XPaths.  Maybe we can keep this approach until that day?

The fix could also be applied to the ModelChange XPath validation:  check to ensure that no structural changes were present, and if not, attempt to validate the XPaths.

The final change is that I changed the text 'could be invalid' to 'generated warnings', because the former seems a little hyperbolic, and many warnings *cannot* make the file invalid; they indicate something else might be wrong.  I feel like the user should be able to peruse the warnings as they are and decide for themselves whether the warnings indicate there might be something invalid present.  (This is particularly true for SBML:  *No* warning indicates that the file may be invalid.  All invalid models are caught with errors, not warnings.  If you prefer the 'may be invalid' wording in general, I request that you leave it for SBML, at least.)

The design of collecting warnings by type could be also applied to errors, but errors are more serious, and always need to all be fixed.  The specific messages indicating where each error comes from I feel are worth the extra potential wordiness of some design error resulting in walls of text.