Fixes the temp-biomodels github actions failure.
The problem here stemmed from pyomexmeta failing to read the XML for some generated files.  We can work on that separately, but in the meantime, this wraps that call in a try/catch block that simply reports a failure to read the file instead of halting.

In addition, this fixes an issue I found on my own runs where UTF-8 characters were confusing the writer:  we now explicitly tell the writer that we are using utf-8.