Add tests; fix bugs.
Several tests added that check XPath validation in various contexts in the presence and absence of a 'structural' model change.  To make things work, the model_etree had to be passed around a lot, and created earlier than it had been.  But everything else mostly worked out of the box.