feat(cellml): add flag to skip over non observable variables in cellml files
add a flag to  skips over variables in the cellML that have been labeled as `"public_interface"="in"` or `"private_interface"="in"` as these are not exposed by OpenCor after simulation. No information is lost from the reports, as these variables have a connection to another component that contains the same variable, with the public interface set to "out". Accessing the variable through that component works fine.