Fix test.
I believe that the errors we now get are on the same order as the ones we used to get, but just different.  "XML content is not well-formed" is pretty similar to "Missing XML declaration"; the file itself is not actually XML at all, so either work for that situation.